requirements.txt
PyQt6==6.6.1
sounddevice==0.4.6
numpy==1.26.3
pandas==2.2.0
requests==2.31.0
python-dotenv==1.0.1 